Move status projection cache into runtime context
